Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
pydantic
python-dotenv
requests
aiofiles
python-magic
google-genai
motor